Move command exec logic into their own classes
